refactor : postskeleton 명칭 변경
